ProjectConfigurationSupport moved to private package, just effective configuration exported.
