+++ extide/gradle/src/org/netbeans/modules/gradle/actions/ActionToTaskUtils.java
@@ -29,7 +29,7 @@ import org.netbeans.api.project.Project;
 import org.netbeans.modules.gradle.api.execute.GradleExecConfiguration;
 import org.openide.util.Lookup;
 import org.netbeans.modules.gradle.execute.ConfigurableActionProvider;
-import org.netbeans.modules.gradle.spi.actions.ProjectConfigurationSupport;
+import org.netbeans.modules.gradle.execute.ProjectConfigurationSupport;
 import org.netbeans.spi.project.ProjectConfigurationProvider;
 
 /**
