+++ extide/gradle/src/org/netbeans/modules/gradle/api/execute/RunConfig.java
@@ -21,6 +21,7 @@ package org.netbeans.modules.gradle.api.execute;
 
 import java.util.Collections;
 import java.util.Set;
+import org.netbeans.api.annotations.common.CheckForNull;
 import org.netbeans.api.project.Project;
 
 /**
@@ -83,9 +84,11 @@ public final class RunConfig {
 
     /**
      * Selected executable configuration. Should be never {@code null}, as the default
-     * configuration always exists.
-     * @return configuration.
+     * configuration always exists. May be {@code null}; in that case the active configuration
+     * should be used.
+     * @return configuration or {@code null} to use the active one.
      */
+    @CheckForNull
     public GradleExecConfiguration getExecConfig() {
         return execConfig;
     }
