+++ extide/gradle/src/org/netbeans/modules/gradle/execute/ProjectConfigurationSupport.java
@@ -16,9 +16,8 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.netbeans.modules.gradle.spi.actions;
+package org.netbeans.modules.gradle.execute;
 
-import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.function.Supplier;
